pyproject.toml
[tool.poetry]
name = "thesis-app"
version = "0.1.0"
description = ""
authors = ["Daniel Engbert <d.engbert@student.vu.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.1"
pytorch-lightning = "^2.2.1"
tensorboard = "^2.16.2"
transformers = "^4.38.2"
huggingface-hub = "^0.21.4"
accelerate = "^0.29.2"
bitsandbytes = "^0.43.1"
langchain = "^0.1.12"
wandb = "^0.16.4"
datasets = "^2.18.0"
evaluate = "^0.4.1"
scikit-learn = "^1.4.1.post1"
notebook = "^7.1.1"
openai = "^1.14.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"
matplotlib = "^3.8.3"
pydantic = "^2.6.4"
numpy = "^1.26.4"
seaborn = "^0.13.2"
haikunator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
mypy = "^1.9.0"
nbstripout = "^0.7.1"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
# Global mypy configurations can be added here
exclude = ["repos", "datasets"]

[[tool.mypy.overrides]]
module = "transformers.*"
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "huggingface_hub.*"
ignore_missing_imports = true
follow_imports = "skip"

[tool.ruff]
exclude = ["repos", "datasets"]

[tool.ruff.lint]
# https://docs.astral.sh/ruff/settings/#lint_ignore-init-module-imports
ignore-init-module-imports = true
# bad for pandas https://github.com/astral-sh/ruff/issues/1852
ignore = ["E712"]